feat(cache): add `Subscribable`
