+++ .changeset/twelve-apricots-decide.md
@@ -1,0 +1,5 @@
+---
+'@suspensive/cache': minor
+---
+
+feat(cache): add `Subscribable`
